Using layout generator with run sequence in mapper
Now using the layout generator in the run sequence dgraph to
generate an optimal layout using data dependency gleaned from
the run sequence.

Also now using doubles instead of floats.

+++ src/Superstructure/Mapper/include/ESMCI_Mapper.h
@@ -23,7 +23,7 @@ namespace ESMCI{
       } MapperOptAlg;
       Mapper(ESMCI::VM &vm); 
       Mapper(ESMCI::VM &vm, const std::string &rseq_fname); 
-      void set_comp_info(const std::vector<MapperUtil::CompInfo<float> > &comp_infos);
+      void set_comp_info(const std::vector<MapperUtil::CompInfo<double> > &comp_infos);
       void add_constraint(const MapperConstraint &constraint);
       void rem_constraint(const MapperConstraint &constraint);
       void add_opt_method(MapperOptAlg opt_alg);
@@ -31,7 +31,7 @@ namespace ESMCI{
       void add_opt_method(const std::vector<MapperOptAlg> &opt_algos);
       bool optimize(std::vector<int> &opt_npets,
                     std::vector<std::pair<int, int> > &opt_pet_ranges,
-                    float &opt_wtime);
+                    double &opt_wtime);
       ~Mapper();
     private:
       ESMCI::VM &vm_;
@@ -39,14 +39,14 @@ namespace ESMCI{
       bool use_rseq_dgraph_dep_;
       const int DEFAULT_LBAL_MAX_ITERS = 10;
       int lbal_max_iters_;
-      std::vector<MapperUtil::CompInfo<float> > comp_infos_;
-      MapperUtil::CompInfoStore<float> *comp_info_store_;
-      MapperUtil::LoadBalancer<float> lb_;
+      std::vector<MapperUtil::CompInfo<double> > comp_infos_;
+      MapperUtil::CompInfoStore<double> *comp_info_store_;
+      MapperUtil::LoadBalancer<double> lb_;
       std::string rseq_fname_;
       MapperUtil::RunSeqDGraph rseq_dgraph_;
 
       void get_rseq_opt_layouts(
-        std::vector<std::vector<MapperUtil::CompInfo<float> > > &opt_layouts);
+        std::vector<std::vector<MapperUtil::CompInfo<double> > > &opt_layouts);
   };
 
 } //namespace ESMCI
